--- apps/charterafrica/src/components/OpportunityCardList/index.js
@@ -1,3 +1,0 @@
-import OpportunityCardList from "./OpportunityCardList";
-
-export default OpportunityCardList;
