+++ apps/charterafrica/src/components/OpportunityCards/OpportunityCardGrid.snap.js
