+++ apps/charterafrica/src/components/OpportunityCards/index.js
@@ -1,0 +1,3 @@
+import OpportunityCards from "./OpportunityCards";
+
+export default OpportunityCards;
